+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.6
+
+- Added Clusterable default value isCluster = false
+
+## 1.0.5
+
+- Added Clusterable constructor
+
+## 1.0.4
+
+- Removed dependency on meta package
+
 ## 1.0.3
 
 - Exporting base_cluster.dart